requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
curlify==2.2.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.1
django-analytical==3.1.0
django-cors-headers==4.4.0
djangorestframework==3.15.1
facebook_business==20.0.0
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
psycopg2-binary==2.9.9
pycountry==24.6.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
stripe==9.7.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
